+++ core-ng-test/src/main/java/core/framework/test/redis/MockRedisSortedSet.java
@@ -20,11 +20,11 @@ public class MockRedisSortedSet implements RedisSortedSet {
 
     @Override
     public boolean add(String key, String value, long score, boolean onlyIfAbsent) {
-        var sortedSet = store.putIfAbsent(key, new HashMap<>()).sortedSet();
+        var map = store.putIfAbsent(key, new HashMap<>()).sortedSet();
         if (onlyIfAbsent) {
-            return sortedSet.putIfAbsent(value, score) == null;
+            return map.putIfAbsent(value, score) == null;
         } else {
-            sortedSet.put(value, score);
+            map.put(value, score);
             return true;
         }
     }
@@ -33,13 +33,13 @@ public class MockRedisSortedSet implements RedisSortedSet {
     public Map<String, Long> range(String key, long start, long stop) {
         var value = store.get(key);
         if (value == null) return Map.of();
-        Map<String, Long> sortedSet = value.sortedSet();
-        int size = sortedSet.size();
+        Map<String, Long> map = value.sortedSet();
+        int size = map.size();
         int startIndex = start < 0 ? 0 : (int) start;
         if (startIndex > size) startIndex = size;
         int endIndex = stop < 0 ? (int) stop + size : (int) stop;
         if (endIndex >= size) endIndex = size - 1;
-        return sortedSet.entrySet().stream()
+        return map.entrySet().stream()
                 .sorted(Entry.comparingByValue())
                 .skip(startIndex)
                 .limit(endIndex - startIndex + 1)
@@ -50,12 +50,12 @@ public class MockRedisSortedSet implements RedisSortedSet {
     public Map<String, Long> popByScore(String key, long minScore, long maxScore, long limit) {
         MockRedisStore.Value value = store.get(key);
         if (value == null) return Map.of();
-        var sortedSet = value.sortedSet();
-        return sortedSet.entrySet().stream()
+        var map = value.sortedSet();
+        return map.entrySet().stream()
                 .filter(entry -> entry.getValue() >= minScore && entry.getValue() <= maxScore)
                 .sorted(Entry.comparingByValue())
                 .limit(limit)
-                .peek(entry -> sortedSet.remove(entry.getKey()))
+                .peek(entry -> map.remove(entry.getKey()))
                 .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue, (key1, key2) -> key2, LinkedHashMap::new));
     }
 }