+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Change log
-### 7.6.1 (10/15/2020 - )
+### 7.6.1 (10/15/2020 - 10/29/2020)
 * http: httpClient does not read response body if status code is 204 
 * executor: improve warning for task rejection/cancellation during shutdown
 * db: added Repository.insertIgnore(entity) to handle duplicated key / constraint violation cases without catching UncheckedSQLException